Finished up
The rest of it should be working. The HelloWorld.vue component was removed since we were not using it.

The specific links have all been fixed such that when they display specific data, the id will be shown in the URL and the data will go along with the router pushing and popping. This system replaced clickId. Furthermore, upon a refresh or history change, sometimes the ID would turn into a string, but that is fixed by casting it into a number at use.