Front End Connection
So I got the front end all connected to the back end. Therefore, the project is now at the same status as it was prior but with a functional database. Now we just need to make the ability to add or delete data from the site itself.

Also, I figured out how to have global functions with Vue prototypes. Those proved invaluable since I was able to reduce a lot of the redundancy between the pages.