Student pages done
I added the functionality to add new students and to delete students (which also has to delete all of the classes that they were registered for). This included modifying the back end a bit, since although the model that we were programmatically loading from consisted of all strings, the json models that were loaded from had all ids as numbers.

Other than that, this update was pretty straight forward. In terms of lines, the majority of it was adding to the HTML for the StudentList.vue and making the corresponding method, which had to verify that each form was filled in before proceeding to submit to the api.